Key rotation code re-struture
Restructure of code additions for key rotation support. This change
moves key rotation code and other small functions to funcs.go,
while rotate.go has been renamed to keyregistry and contains just
the key registry code.
